Move Navbar.jsx to a separate folder inside components folder.
